initial draft for C API
